moved to correct package and now compiles
