+++ wrapper/src/main/java/software/amazon/jdbc/plugin/encryption/logging/AuditLogger.java
@@ -15,7 +15,7 @@
  */
 
 
-package software.amazon.jdbc.logging;
+package software.amazon.jdbc.plugin.encryption.logging;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
