+++ wrapper/src/main/java/software/amazon/jdbc/plugin/encryption/wrapper/EncryptingPreparedStatement.java
@@ -17,11 +17,11 @@
 
 package software.amazon.jdbc.plugin.encryption.wrapper;
 
-import software.amazon.jdbc.metadata.MetadataManager;
+import software.amazon.jdbc.plugin.encryption.metadata.MetadataManager;
 import software.amazon.jdbc.plugin.encryption.model.ColumnEncryptionConfig;
 import software.amazon.jdbc.plugin.encryption.key.KeyManager;
-import software.amazon.jdbc.service.EncryptionService;
-import software.amazon.jdbc.sql.SqlAnalysisService;
+import software.amazon.jdbc.plugin.encryption.service.EncryptionService;
+import software.amazon.jdbc.plugin.encryption.sql.SqlAnalysisService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
